LinearMaps.LinearMap

TPSChem.ActiveSpaceSolvers.FCI.FCIAnsatz

TPSChem.ActiveSpaceSolvers.FCI.FCIAnsatz

TPSChem.ActiveSpaceSolvers.Solution

TPSChem.ActiveSpaceSolvers.SolverSettings

TPSChem.ActiveSpaceSolvers.SolverSettings

TPSChem.BlockDavidson.LinOpMat

Base.copy!

TPSChem.ActiveSpaceSolvers.FCI.apply_S2_matrix

TPSChem.ActiveSpaceSolvers.FCI.apply_annihilation!

TPSChem.ActiveSpaceSolvers.FCI.apply_creation!

TPSChem.ActiveSpaceSolvers.FCI.build_S2_matrix

TPSChem.ActiveSpaceSolvers.FCI.calc_linear_index

TPSChem.ActiveSpaceSolvers.FCI.calc_linear_index!

TPSChem.ActiveSpaceSolvers.FCI.calc_linear_index!

TPSChem.ActiveSpaceSolvers.FCI.calc_nchk

TPSChem.ActiveSpaceSolvers.FCI.compute_1rdm

TPSChem.ActiveSpaceSolvers.FCI.compute_AA

TPSChem.ActiveSpaceSolvers.FCI.compute_AAa

TPSChem.ActiveSpaceSolvers.FCI.compute_AB

TPSChem.ActiveSpaceSolvers.FCI.compute_ABa

TPSChem.ActiveSpaceSolvers.FCI.compute_ABb

TPSChem.ActiveSpaceSolvers.FCI.compute_Aa

TPSChem.ActiveSpaceSolvers.FCI.compute_Ab

TPSChem.ActiveSpaceSolvers.FCI.compute_S2_expval

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms2

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms_full

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms_full!

TPSChem.ActiveSpaceSolvers.FCI.compute_creation

TPSChem.ActiveSpaceSolvers.FCI.compute_fock_diagonal

TPSChem.ActiveSpaceSolvers.FCI.compute_rdm1_rdm2

TPSChem.ActiveSpaceSolvers.FCI.compute_ss_terms2

TPSChem.ActiveSpaceSolvers.FCI.fill_ca_lookup

TPSChem.ActiveSpaceSolvers.FCI.fill_ca_lookup2

TPSChem.ActiveSpaceSolvers.FCI.fill_ca_lookup3

TPSChem.ActiveSpaceSolvers.FCI.get_map

TPSChem.ActiveSpaceSolvers.FCI.get_map

TPSChem.ActiveSpaceSolvers.FCI.get_nchk

TPSChem.ActiveSpaceSolvers.FCI.get_unoccupied

TPSChem.ActiveSpaceSolvers.FCI.get_unoccupied!

TPSChem.ActiveSpaceSolvers.FCI.precompute_spin_diag_terms

TPSChem.ActiveSpaceSolvers.FCI.reset!

TPSChem.ActiveSpaceSolvers.FCI.run_fci

TPSChem.ActiveSpaceSolvers.FCI.svd_state

TPSChem.ActiveSpaceSolvers.apply_S2_matrix

TPSChem.ActiveSpaceSolvers.apply_sminus

TPSChem.ActiveSpaceSolvers.apply_splus

TPSChem.ActiveSpaceSolvers.build_H_matrix

TPSChem.ActiveSpaceSolvers.build_S2_matrix

TPSChem.ActiveSpaceSolvers.compute_1rdm

TPSChem.ActiveSpaceSolvers.compute_1rdm_2rdm

TPSChem.ActiveSpaceSolvers.compute_operator_c_a

TPSChem.ActiveSpaceSolvers.compute_operator_c_b

TPSChem.ActiveSpaceSolvers.compute_operator_ca_aa

TPSChem.ActiveSpaceSolvers.compute_operator_ca_ab

TPSChem.ActiveSpaceSolvers.compute_operator_ca_bb

TPSChem.ActiveSpaceSolvers.compute_operator_cc_aa

TPSChem.ActiveSpaceSolvers.compute_operator_cc_ab

TPSChem.ActiveSpaceSolvers.compute_operator_cc_bb

TPSChem.ActiveSpaceSolvers.compute_operator_cca_aaa

TPSChem.ActiveSpaceSolvers.compute_operator_cca_aba

TPSChem.ActiveSpaceSolvers.compute_operator_cca_abb

TPSChem.ActiveSpaceSolvers.compute_operator_cca_bbb

TPSChem.ActiveSpaceSolvers.compute_s2

TPSChem.ActiveSpaceSolvers.solve

TPSChem.ActiveSpaceSolvers.svd_state

ActiveSpaceSolvers

TPSChem.ActiveSpaceSolvers.Solution — Type

This type contains both the Ansatz and the results.

- ansatz::A
- energies::Vector{T}
- vectors::Matrix{T}

TPSChem.ActiveSpaceSolvers.SolverSettings — Type

This type contains the solver settings information needed to solve the problem.

- nroots::Int
- tol::Float64
- maxiter::Int
- verbose::Int
- package::String ["arpack", "davidson"]

TPSChem.ActiveSpaceSolvers.SolverSettings — Method

SolverSettings(;nroots=1, tol=1e-8, maxiter=100, verbose=0, package="arpack")

Default value constructor

TPSChem.ActiveSpaceSolvers.solve — Method

solve(ints::InCoreInts{T}, ansatz::A, S::SolverSettings; v0=nothing) where {T, A<:Ansatz}

Get the energies and eigenstates (stored as a Solution{A,T} type), for the Hamiltonian (defined by ints) with the wavefunction approximated by the ansatz (defined by ansatz), and passing the solver settings (defined by S) to the solver.

Arguments

ints: Integrals

ansatz: Subtype of the abstract type, Ansatz

S: SolverSettings

v0: initial guess (optional). If provided with "davidson", it needs to have number of columns equal to number of roots sought. If provided with "arpack", it can only be one vector.

LinearMaps.LinearMap — Method

LinearMap(ints, prb::FCIAnsatz)

Get LinearMap with takes a vector and returns action of H on that vector

Arguments

ints: InCoreInts object

prb: FCIAnsatz object

TPSChem.ActiveSpaceSolvers.FCI.FCIAnsatz — Type

Type containing all the metadata needed to define a FCI problem

no::Int  # number of orbitals
na::Int  # number of alpha
nb::Int  # number of beta
dima::Int 
dimb::Int 
dim::Int
converged::Bool
restarted::Bool
iteration::Int
algorithm::String   #  options: direct/davidson
n_roots::Int

TPSChem.ActiveSpaceSolvers.FCI.FCIAnsatz — Method

FCIAnsatz(no, na, nb)

Constructor

Arguments

no: Number of spatial orbitals

na: Number of α electrons

nb: Number of β electrons

TPSChem.BlockDavidson.LinOpMat — Method

LinOpMat(ints, prb::FCIAnsatz)

Get LinearMap with takes a vector and returns action of H on that vector

Arguments

ints: InCoreInts object

prb: FCIAnsatz object

Base.copy! — Method

Base.copy!(c1::DeterminantString,c2::DeterminantString)

copy c2 into c1

TPSChem.ActiveSpaceSolvers.FCI.apply_S2_matrix — Method

apply_S2_matrix(prb::FCIAnsatz, v::AbstractArray{T}) where {T}

prb: FCIAnsatz just defines the current CI ansatz (i.e., fock sector)

TPSChem.ActiveSpaceSolvers.FCI.apply_annihilation! — Method

apply_annihilation!(c::DeterminantString, orb_index::Int)

Apply an annihilation operator to c corresponding to orbital orb_index

TPSChem.ActiveSpaceSolvers.FCI.apply_creation! — Method

apply_creation!(c::DeterminantString, orb_index::Int)

Apply a creation operator to c corresponding to orbital orb_index

TPSChem.ActiveSpaceSolvers.FCI.build_S2_matrix — Method

build_s2(prb::FCIAnsatz)

prb: FCIAnsatz just defines the current CI ansatz (i.e., fock sector)

TPSChem.ActiveSpaceSolvers.FCI.calc_linear_index! — Method

calc_linear_index!(c::DeterminantString, binomcoeff::Array{Int,2})

Calculate the linear index, passing in binomial coefficient matrix makes it much faster

TPSChem.ActiveSpaceSolvers.FCI.calc_linear_index! — Method

calc_linear_index!(c::DeterminantString)

Calculate the linear index

TPSChem.ActiveSpaceSolvers.FCI.calc_linear_index — Method

calc_linear_index(c::DeterminantString)

Return linear index for lexically ordered __config DeterminantString

TPSChem.ActiveSpaceSolvers.FCI.calc_nchk — Method

calc_nchk(n::Integer,k::Integer)

Calculates binomial coefficient: n choose k

TPSChem.ActiveSpaceSolvers.FCI.compute_1rdm — Method

TPSChem.ActiveSpaceSolvers.FCI.compute_AA — Method

compute_AA(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute represntation of 2creation operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta)

TPSChem.ActiveSpaceSolvers.FCI.compute_AAa — Method

compute_AAa(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute represntation of 2creation operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta)

TPSChem.ActiveSpaceSolvers.FCI.compute_AB — Method

compute_AB(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix)

Compute representation of p'q' operators between states bra_v and ket_v, where p is alpha and q is beta.

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

G(pqst) = v(IJs) <IJ|p'q'|KL> v(KLt) 
        = v(IJs) <J|<I|p'q'|K>|L> v(KLt)
        = v(IJs) <J|<I|p'|K>q'|L> v(KLt) (-1)^N(K)
        = v(IJs) <I|p'|K> <J|q'|L> v(KLt) (-1)^N(K)
    
        this can be vectorized if needed. for now:
        loop over p,q
            loop over K,L
                sign = -1^N(K)
                find indices of I,J
                G(pq,st) += v(IJ,s) v(KL,t) * sign 
            end
        end

TPSChem.ActiveSpaceSolvers.FCI.compute_ABa — Method

compute_ABa(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute represntation of 2creation operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta) giving ABa or BAb

TPSChem.ActiveSpaceSolvers.FCI.compute_ABb — Method

compute_ABb(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute represntation of 2creation operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta) giving ABa or BAb

TPSChem.ActiveSpaceSolvers.FCI.compute_Aa — Method

compute_Aa(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute representation of a'a operators between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta)

TPSChem.ActiveSpaceSolvers.FCI.compute_Ab — Method

compute_Ab(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix)

Compute representation of p'q operators between states bra_v and ket_v, where p is alpha and q is beta.

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

G(pqst) = v(IJs) <IJ|p'q|KL> v(KLt) 
        = v(IJs) <J|<I|p'q|K>|L> v(KLt)
        = v(IJs) <J|<I|p'|K>q|L> v(KLt) (-1)^N(K)
        = v(IJs) <I|p'|K> <J|q|L> v(KLt) (-1)^N(K)
    
        this can be vectorized if needed. for now:
        loop over p,q
            loop over K,L
                sign = -1^N(K)
                find indices of I,J
                G(pq,st) += v(IJ,s) v(KL,t) * sign 
            end
        end

TPSChem.ActiveSpaceSolvers.FCI.compute_S2_expval — Method

compute_S2_expval(prb::FCIAnsatz)

prb: FCIAnsatz just defines the current CI ansatz (i.e., fock sector)

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms — Method

compute_ab_terms(v, H, P::FCIAnsatz, ket_a_lookup, ket_b_lookup)

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms — Method

compute_ab_terms(v, H, P::FCIAnsatz)

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms2 — Method

compute_ab_terms2(v, H, P::FCIAnsatz, 
                      ket_a_lookup, ket_b_lookup)

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms_full! — Method

compute_ab_terms_full!(H, P::FCIAnsatz, Hmat)

TPSChem.ActiveSpaceSolvers.FCI.compute_ab_terms_full — Method

compute_ab_terms_full(H, P::FCIAnsatz)

TPSChem.ActiveSpaceSolvers.FCI.compute_creation — Method

compute_creation(no::Integer, bra_na, bra_nb, ket_na, ket_nb, bra_v::Matrix, ket_v::Matrix, spin_case)

Compute an creation operator between states bra_v and ket_v

Arguments

no: number of orbitals

bra_na: number of alpha electrons in bra

bra_nb: number of beta electrons in bra

ket_na: number of alpha electrons in ket

ket_nb: number of beta electrons in ket

bra_v: basis vectors in bra

ket_v: basis vectors ket

spin_case: either (alpha|beta)

TPSChem.ActiveSpaceSolvers.FCI.compute_fock_diagonal — Method

compute_fock_diagonal!(H, P::FCIAnsatz, e_mf::Float64)

TPSChem.ActiveSpaceSolvers.FCI.compute_rdm1_rdm2 — Method

compute_rdm1_rdm2(P::FCIAnsatz, vec_l::Vector{T}, vec_r::Vector{T}) where T

Compute the 1 and 2 rdm's, in pyscf order

TPSChem.ActiveSpaceSolvers.FCI.compute_ss_terms2 — Method

compute_aa_terms2(v, H, P::FCIAnsatz, 
                      ket_a_lookup)

TPSChem.ActiveSpaceSolvers.FCI.fill_ca_lookup — Method

fill_ca_lookup(c::DeterminantString)

Create an index table relating each DeterminantString with all ia substitutions i.e., calookup[Ka][c(p) + a(p)*np] = La

TPSChem.ActiveSpaceSolvers.FCI.fill_ca_lookup2 — Method

fill_ca_lookup2(c::DeterminantString)

Create an index table relating each DeterminantString with all ia substitutions i.e., ca_lookup[Ka,p,q] = sign*La

<L|p'q|K> = sign

TPSChem.ActiveSpaceSolvers.FCI.fill_ca_lookup3 — Method

fill_ca_lookup3(c::DeterminantString)

Create an index table relating each DeterminantString with all ia substitutions i.e., ca_lookup[Ka,p,q] = (sign,La)

<L|p'q|K> = sign

TPSChem.ActiveSpaceSolvers.FCI.get_map — Method

get_map(ham, prb::FCIAnsatz, HdiagA, HdiagB)

Assumes you've already computed the spin diagonal components

TPSChem.ActiveSpaceSolvers.FCI.get_map — Method

get_map(ham, prb::FCIAnsatz)

Get LinearMap with takes a vector and returns action of H on that vector

TPSChem.ActiveSpaceSolvers.FCI.get_nchk — Method

get_nchk(n::Integer,k::Integer)

Looks up binomial coefficient from a precomputed table: n choose k

TPSChem.ActiveSpaceSolvers.FCI.get_unoccupied! — Method

get_unoccupied(c::DeterminantString)

get list of orbitals that are unoccupied

TPSChem.ActiveSpaceSolvers.FCI.get_unoccupied — Method

get_unoccupied(c::DeterminantString)

get list of orbitals that are unoccupied

TPSChem.ActiveSpaceSolvers.FCI.precompute_spin_diag_terms — Method

precompute_spin_diag_terms(H, P::FCIAnsatz, e)

TPSChem.ActiveSpaceSolvers.FCI.reset! — Method

reset!(c::DeterminantString)

Reset the DeterminantString to the first config

TPSChem.ActiveSpaceSolvers.FCI.run_fci — Method

run_fci(ints, prb::FCIAnsatz)

input: ints is a struct containing 0, 2, and 4 dimensional tensors

h0: energy shift

h1: 1 electron integrals

h2: 2 electron integrals (chemists notation)

prb: FCIAnsatz just defines the current CI ansatz (i.e., fock sector)

ints is simply an InCoreInts object from TPSChem

TPSChem.ActiveSpaceSolvers.FCI.svd_state — Method

svd_state(prb::FCIAnsatz)

Do an SVD of the FCI vector partitioned into clusters with (norbs1 | norbs2) where the orbitals are assumed to be ordered for cluster 1| cluster 2 haveing norbs1 and norbs2, respectively.

prb: FCIAnsatz just defines the current CI ansatz (i.e., fock sector)

norbs1:number of orbitals in left cluster

norbs2:number of orbitals in right cluster

svd_thresh: the threshold below which the states will be discarded

TPSChem.ActiveSpaceSolvers.apply_S2_matrix — Method

build_S2_matrix(P::FCIAnsatz)

Build the S2 matrix in the Slater Determinant Basis specified by P

TPSChem.ActiveSpaceSolvers.apply_sminus — Method

TPSChem.ActiveSpaceSolvers.apply_splus — Method

TPSChem.ActiveSpaceSolvers.build_H_matrix — Method

build_H_matrix(ints, P::FCIAnsatz)

Build the Hamiltonian defined by ints in the Slater Determinant Basis specified by P

TPSChem.ActiveSpaceSolvers.build_S2_matrix — Method

build_S2_matrix(P::FCIAnsatz)

Build the S2 matrix in the Slater Determinant Basis specified by P

TPSChem.ActiveSpaceSolvers.compute_1rdm — Method

compute_1rdm(sol::Solution{FCIAnsatz,T}; root=1) where {T}

TPSChem.ActiveSpaceSolvers.compute_1rdm_2rdm — Method

compute_2rdm(sol::Solution{A,T}; root=1) where {A,T}

TPSChem.ActiveSpaceSolvers.compute_operator_c_a — Method

compute_operator_c_a(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a operator between states bra_v and ket_v for alpha

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_c_b — Method

compute_operator_c_b(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a operator between states bra_v and ket_v for beta

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_ca_aa — Method

compute_operator_ca_aa(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a operators between states bra_v and ket_v for alpha-alpha

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_ca_ab — Method

compute_operator_ca_ab(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a operators between states bra_v and ket_v for alpha-beta

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_ca_bb — Method

compute_operator_ca_bb(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a operators between states bra_v and ket_v for beta-beta

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_cc_aa — Method

compute_operator_cc_aa(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a' operators between states bra_v and ket_v for alpha-alpha

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_cc_ab — Method

compute_operator_cc_ab(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a' operators between states bra_v and ket_v for alpha-beta

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_cc_bb — Method

compute_operator_cc_aa(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a' operators between states bra_v and ket_v for beta-beta

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_cca_aaa — Method

compute_operator_cca_aaa(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a'a operators between states bra_v and ket_v for alpha-alpha-alpha

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_cca_aba — Method

compute_operator_cca_aba(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a'a operators between states bra_v and ket_v for alpha-beta-alpha

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_cca_abb — Method

compute_operator_cca_abb(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a'a operators between states bra_v and ket_v for alpha-beta-beta

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_operator_cca_bbb — Method

compute_operator_cca_bbb(bra::Solution{FCIAnsatz,T}, ket::Solution{FCIAnsatz,T}) where {T}

Compute representation of a'a'a operators between states bra_v and ket_v for beta-beta-beta

Arguments

bra: solutions for the left hand side

ket: solutions for the right hand side

TPSChem.ActiveSpaceSolvers.compute_s2 — Method

ActiveSpaceSolvers.compute_s2(sol::Solution)

Compute the <S^2> expectation values for each state in sol

TPSChem.ActiveSpaceSolvers.svd_state — Method

svd_state(sol::Solution{FCIAnsatz,T},norbs1,norbs2,svd_thresh; root=1) where T

Do an SVD of the FCI vector partitioned into clusters with (norbs1 | norbs2) where the orbitals are assumed to be ordered for cluster 1| cluster 2 haveing norbs1 and norbs2, respectively.

sol: Solution just defines the current CI states

norbs1:number of orbitals in left cluster

norbs2:number of orbitals in right cluster

svd_thresh: the threshold below which the states will be discarded

root: which root to SVD