QCBase

Foundational quantum-chemistry data types shared by every method in the package: atoms, molecules, and molecular-orbital clusters (MOCluster), together with the small generic interface (n_orb, dim_tot, orbital_rotation, ...) that the other submodules specialize.

TPSChem.QCBase.Atom

TPSChem.QCBase.MOCluster

TPSChem.QCBase.Molecule

Base.length

TPSChem.QCBase.dim_tot

TPSChem.QCBase.dim_tot

TPSChem.QCBase.possible_focksectors

Documentation

TPSChem.QCBase.Atom — Type

id::Integer             #index of atom in the molecule
symbol::String          #Atomic ID (E.g. H, He, ...)
xyz::Array{Float64,1}   #list of XYZ coordinates

Simply an Atom

TPSChem.QCBase.MOCluster — Type

idx::Int16
orb_list::Vector{Int16}

TPSChem.QCBase.Molecule — Type

charge::Integer         #overall charge on molecule
multiplicity::Integer   #2S+1
atoms::Vector{Atom}     #Vector of `Atoms`
basis::String           #Basis set

Molecule essentially as a Vector of atoms, number of electrons and basis set

Base.length — Method

length(c::MOCluster)

Return number of orbitals in MOCluster

TPSChem.QCBase.dim_tot — Method

dim_tot(c::MOCluster, na, nb)

Return dimension of hilbert space spanned by number of orbitals in MOCluster with na and nb number of alpha/beta electrons.

TPSChem.QCBase.dim_tot — Method

dim_tot(c::MOCluster)

Return dimension of hilbert space spanned by number of orbitals in MOCluster. This is all sectors

TPSChem.QCBase.possible_focksectors — Method

possible_focksectors(c::MOCluster, delta_elec=nothing)

Get list of possible fock spaces accessible to the MOCluster

delta_elec::Vector{Int}: (nα, nβ, Δ) restricts fock spaces to: (nα,nβ) ± Δ electron transitions