RDM

The RDM submodule provides the reduced-density-matrix types (RDM1, RDM2 and their spin-summed variants ssRDM1 / ssRDM2) together with the orbital gradient, orbital Hessian, and generalized Fock machinery that drive the cluster mean-field orbital optimization.

Note

These are the building-block density matrices and the gradient/Hessian used by CMF orbital optimization. RDMs measured from a converged TPSCI wavefunction are computed separately — see RDMs & Properties.

TPSChem.RDM.Cumulant2

TPSChem.RDM.RDM1

TPSChem.RDM.RDM1

TPSChem.RDM.RDM2

TPSChem.RDM.ssRDM1

TPSChem.RDM.ssRDM2

TPSChem.InCoreIntegrals.subset

TPSChem.QCBase.compute_energy

TPSChem.QCBase.compute_energy

TPSChem.QCBase.compute_energy

TPSChem.RDM.build_generalised_fock

TPSChem.RDM.build_generalised_fock

TPSChem.RDM.build_orbital_gradient

TPSChem.RDM.build_orbital_gradient

TPSChem.RDM.build_orbital_hessian

TPSChem.RDM.build_orbital_hessian

TPSChem.RDM.compute_fock

TPSChem.RDM.pack_gradient

TPSChem.RDM.pack_hessian

TPSChem.RDM.unpack_gradient

TPSChem.RDM.unpack_hessian

Documentation

TPSChem.RDM.Cumulant2 — Type

aa::Array{T,2}  # αα
ab::Array{T,2}  # αβ 
bb::Array{T,2}  # ββ

2 Cumulant

TPSChem.RDM.RDM1 — Type

a::Array{T,2}   # α
b::Array{T,2}   # β

1 particle RDM type

TPSChem.RDM.RDM1 — Method

RDM1(rdm::RDM2)

Integrate a 2rdm to get the 1rdm. We assume that d is stored, d[1,1,2,2] such that <p'q'rs> is D[p,s,q,r] Also, we will agree with pyscf, and use the following normalization: tr(D) = N(N-1)

TPSChem.RDM.RDM2 — Type

aa::Array{T,2}  # αα
ab::Array{T,2}  # αβ 
bb::Array{T,2}  # ββ

2 particle RDM type

TPSChem.RDM.ssRDM1 — Type

rdm::Array{T,2}

Spin-summed 1-particle RDM

TPSChem.RDM.ssRDM2 — Type

rdm::Array{T,4}

Spin-summed 2-particle RDM

TPSChem.InCoreIntegrals.subset — Method

subset(ints::InCoreInts, list, rmd1a, rdm1b)

Extract a subset of integrals acting on orbitals in list, returned as InCoreInts type and contract a 1rdm to give effectve 1 body interaction

Arguments

ints::InCoreInts: Integrals for full system

list: list of orbital indices in subset

rdm1a: 1RDM for embedding α density to make CASCI hamiltonian

rdm1b: 1RDM for embedding β density to make CASCI hamiltonian

TPSChem.QCBase.compute_energy — Method

compute_energy(ints::InCoreInts{T}, d1::RDM1{T}, d2::RDM2{T}) where T

Return energy defined by rdm1 and rdm2.

Arguments

ints: InCoreInts object

d1: 1 particle reduced density matrix

d2: 2 particle reduced density matrix

TPSChem.QCBase.compute_energy — Method

compute_energy(ints::InCoreInts{T}, rdm1::RDM1{T}) where T

Return energy defined by rdm1.

Arguments

ints: InCoreInts object

rdm1: 1 particle reduced density matrix

TPSChem.QCBase.compute_energy — Method

compute_energy(ints::InCoreInts{T}, d1::ssRDM1{T}, d2::ssRDM2{T}) where T

Return energy defined by spin-summed rdm1 and rdm2.

Arguments

ints: InCoreInts object

d1: 1 particle reduced density matrix

d2: 2 particle reduced density matrix

TPSChem.RDM.build_generalised_fock — Method

build_generalised_fock(ints, rdm1::RDM1{T}, rdm2::RDM2{T})

Build fock-like term for cmf (non spin-summed) f{pq} = h{pq}(d1.a+d1.b) + V_{p,q,r,s}Γ_{t,q,r,s}

Arguments

ints: Integrals

rdm1: Not spin-summed 1RDM

rdm2: Not spin-summed 2RDM

TPSChem.RDM.build_generalised_fock — Method

build_generalised_fock(ints, rdm1::ssRDM1{T}, rdm2::ssRDM2{T})

Build fock-like term f{pq} = h{pq} γ{pq} + V{r,p,s,t} Γ_{r,q,s,t}

Arguments

ints: Integrals

rdm1: Spin summed 1RDM, NxN

rdm2: Spin summed 2RDM, NxNxNxN

TPSChem.RDM.build_orbital_gradient — Method

build_orbital_gradient(ints, rdm1::D1, d2::RDM2)

Build the full orbital rotation hessian g_{pq} = <[H,p'q-q'p]> for p<q

Arguments

ints: Integrals

d1: Not spin-summed

d2: Not spin-summed

TPSChem.RDM.build_orbital_gradient — Method

build_orbital_gradient(ints, rdm1::ssRDM1{T}, rdm2::ssRDM2{T})

Build the full orbital rotation hessian g_{pq} = <[H,p'q-q'p]> for p<q

Arguments

ints: Integrals

rdm1: Spin summed 1RDM, NxN

rdm2: Spin summed 2RDM, NxNxNxN

TPSChem.RDM.build_orbital_hessian — Method

build_orbital_hessian(ints, rdm1::RDM1{T}, rdm2::RDM2{T})

Build the full orbital rotation hessian H_{pq,rs} = <[[H,p'q-q'p], r's-s'r]> (non spin-summed)

Arguments

ints: Integrals

rdm1: Not spin-summed 1RDM

rdm2: Not spin-summed 2RDM

TPSChem.RDM.build_orbital_hessian — Method

build_orbital_hessian(ints, rdm1::ssRDM1{T}, rdm2::ssRDM2{T})

Build the full orbital rotation hessian H_{pq,rs} = <[[H,p'q-q'p], r's-s'r]>

Arguments

ints: Integrals

rdm1: Spin summed 1RDM

rdm2: Spin summed 2RDM

TPSChem.RDM.compute_fock — Method

compute_fock(ints::InCoreInts, rdm1::RDM1)

Compute Fock Matrix

TPSChem.RDM.pack_gradient — Method

pack_gradient(K,norb)

Pack the orbital rotation parameter to upper-triangle

TPSChem.RDM.pack_hessian — Method

pack_hessian(H, norb)

Pack the rank-4 hessian tensor to a symmetric matrix of size n(n-1)/2 × n(n-1)/2.

TPSChem.RDM.unpack_gradient — Method

unpack_gradient(kappa,norb)

Unpack the orbital rotation parameter to full matrix

TPSChem.RDM.unpack_hessian — Method

unpack_hessian(h, norb)

Unpack a packed hessian matrix back to a rank-4 tensor.