States & Configurations

Each method in the package uses a dedicated data representation for the wavefunction. These all rely on hash-table lookups keyed by compact, stack-allocated configuration indices.

Configuration (index) types

FockConfig — how the electrons are distributed over the clusters (per-cluster (nα, nβ)).

ClusterConfig — which local many-body state each cluster is in.

TuckerConfig — a range of local states per cluster (a block).

TransferConfig, OperatorConfig, SparseIndex — bookkeeping for operator application between Fock sectors.

Wavefunction types (each maps configuration indices → coefficients)

TPSCI: TPSCIstate maps FockConfig → ClusterConfig → a vector of coefficients (one per root).

SPT: SPTstate maps FockConfig → TuckerConfig → a compressed Tucker factor for that block.

BSstate — a block-sparse Tucker-form state.

The per-cluster basis and operator data live in ClusterBasis, ClusterSubspace, and ClusterOps.

Index

TPSChem.AbstractState

TPSChem.BSstate

TPSChem.BSstate

TPSChem.BSstate

TPSChem.BSstate

TPSChem.BSstate

TPSChem.ClusterBasis

TPSChem.ClusterBasis

TPSChem.ClusterConfig

TPSChem.ClusterOps

TPSChem.ClusterSubspace

TPSChem.FockConfig

TPSChem.OperatorConfig

TPSChem.SPTstate

TPSChem.SPTstate

TPSChem.SPTstate

TPSChem.SPTstate

TPSChem.SPTstate

TPSChem.SPTstate

TPSChem.SPTstate

TPSChem.SparseIndex

TPSChem.TPSCIstate

TPSChem.TPSCIstate

TPSChem.TPSCIstate

TPSChem.TPSCIstate

TPSChem.TransferConfig

TPSChem.TuckerConfig

Base.:*

Base.:*

Base.:*

Base.:+

Base.:+

Base.:+

Base.:-

Base.:-

Base.:-

Base.Multimedia.display

Base.Multimedia.display

Base.Multimedia.display

Base.convert

Base.convert

Base.convert

Base.convert

Base.length

Base.length

Base.setindex!

Base.size

LinearAlgebra.dot

LinearAlgebra.dot

LinearAlgebra.dot

LinearAlgebra.norm

LinearAlgebra.norm

TPSChem._add_cmf_operator_for_cluster!

TPSChem._compute_cluster_ops_for_cluster

TPSChem.add!

TPSChem.add!

TPSChem.add_config!

TPSChem.add_fockconfig!

TPSChem.add_fockconfig!

TPSChem.add_fockconfig!

TPSChem.add_fockconfig!

TPSChem.add_spin_focksectors

TPSChem.add_spin_focksectors

TPSChem.add_spinfree_2rdm_ops!

TPSChem.clip!

TPSChem.compress

TPSChem.compress_iteratively

TPSChem.compute_cluster_eigenbasis

TPSChem.compute_cluster_eigenbasis_spin

TPSChem.compute_cluster_est_basis

TPSChem.compute_cluster_ops_2rdm

TPSChem.ct_analysis

TPSChem.ct_table

TPSChem.ct_table

TPSChem.dim

TPSChem.dim

TPSChem.expand_each_fock_space!

TPSChem.extract_chosen_root

TPSChem.extract_roots

TPSChem.eye!

TPSChem.eye!

TPSChem.eye!

TPSChem.fold!

TPSChem.form_schmidt_basis

TPSChem.general_ct_analysis

TPSChem.general_ct_table

TPSChem.get_ortho_compliment

TPSChem.get_vector

TPSChem.get_vector

TPSChem.get_vector

TPSChem.get_vector

TPSChem.get_vector

TPSChem.get_vector

TPSChem.get_vector!

TPSChem.is_subset

TPSChem.merge_cluster_bases

TPSChem.mult!

TPSChem.mult!

TPSChem.n_elec

TPSChem.n_elec_a

TPSChem.n_elec_b

TPSChem.nonorth_add

TPSChem.nonorth_add!

TPSChem.nonorth_dot

TPSChem.nonorth_overlap

TPSChem.normalize!

TPSChem.orth!

TPSChem.orth_add

TPSChem.orth_add!

TPSChem.orth_dot

TPSChem.orth_dot

TPSChem.orth_overlap

TPSChem.orthogonalize!

TPSChem.orthonormalize!

TPSChem.orthonormalize!

TPSChem.orthonormalize!

TPSChem.overlap

TPSChem.overlap

TPSChem.possible_spin_focksectors

TPSChem.print_configs

TPSChem.print_fock_occupations

TPSChem.print_fock_occupations

TPSChem.project_into_new_basis

TPSChem.prune_empty_TuckerConfigs!

TPSChem.prune_empty_fock_spaces!

TPSChem.prune_empty_fock_spaces!

TPSChem.prune_empty_fock_spaces!

TPSChem.randomize!

TPSChem.randomize!

TPSChem.randomize!

TPSChem.replace

TPSChem.replace

TPSChem.replace

TPSChem.replace

TPSChem.rotate!

TPSChem.rotate!

TPSChem.scale

TPSChem.scale!

TPSChem.scale!

TPSChem.scale!

TPSChem.scale!

TPSChem.set_vector!

TPSChem.set_vector!

TPSChem.set_vector!

TPSChem.set_vector!

TPSChem.set_vector!

TPSChem.set_vector!

TPSChem.tdm_A

TPSChem.tdm_AA

TPSChem.tdm_AAa

TPSChem.tdm_AB

TPSChem.tdm_ABa

TPSChem.tdm_Aa

TPSChem.tdm_Ab

TPSChem.tdm_H

TPSChem.tdm_S2

TPSChem.tdm_spinfree_2rdm

TPSChem.unfold!

TPSChem.zero!

TPSChem.zero!

TPSChem.zero!

Types

TPSChem.AbstractState — Type

Glue different Sparse Vector State types together

TPSChem.BSstate — Type

Represents a state in an set of abitrary (yet dense) subspaces of a set of FockConfigs.

v[FockConfig][TuckerConfig] => Dense matrix: C(n1n2...nN, R) E.g., used in n-body Tucker

clusters::Vector{MOCluster}
data::OrderedDict{FockConfig,OrderedDict{TuckerConfig,Array}}
p_spaces::Vector{ClusterSubspace}
q_spaces::Vector{ClusterSubspace}

TPSChem.BSstate — Method

BSstate(clusters; T=Float64, R=1)

Constructor creating an empty vector

Arguments

clusters::Vector{MOCluster}

T: Type of data for coefficients

R: Number of roots

Returns

BSstate

TPSChem.BSstate — Method

BSstate(clusters::Vector{MOCluster}, 
         p_spaces::Vector{ClusterSubspace},
         cluster_bases::Vector{ClusterBasis}; R=1) where {T,A}

Constructor needing only prespecified p-spaces. The Q space is defined as the orthogonal complement of p-space within the available basis, specified by cluster_bases.

Arguments

clusters: vector of clusters types

p_spaces: list of pspaces for each cluster

cluster_basis: list of ClusterBasis types - needed to know the dimensions of the q-spaces

R: number of roots

Returns

BSstate

TPSChem.BSstate — Method

function BSstate(v::BSstate{T,N,R}; T=T, R=R) where {T,N,R}

Constructor creating a BSstate with the same basis as v, but with potentially different R and T. Coefficients of new vector are 0.0

Arguments

T: Type of data for coefficients

R: Number of roots

Returns

BSstate

TPSChem.BSstate — Method

BSstate(clusters::Vector{MOCluster}, 
    fconfig::FockConfig{N}, 
    cluster_bases::Vector{ClusterBasis}) where {N}

Constructor using only a single FockConfig. This allows us to turn the CMF state into a BSstate. As such, it chooses the ground state of each cluster in the Fock sector specified by FockConfig to be the P space, and then the Q space is defined as the orthogonal complement of this state within the available basis, specified by cluster_bases.

Arguments

clusters: vector of clusters types

fconfig: starting FockConfig

cluster_basis: list of ClusterBasis types - needed to know the dimensions of the q-spaces

Returns

BSstate

TPSChem.TPSCIstate — Type

clusters::Vector{MOCluster}
data::OrderedDict{FockConfig{N}, OrderedDict{ClusterConfig{N}, Vector{T}}}

This represents an arbitrarily sparse state. E.g., used in TPSCI

TPSChem.TPSCIstate — Method

TPSCIstate(clusters; T=Float64, R=1)

Constructor creating an empty vector

Arguments

clusters::Vector{MOCluster}

T: Type of data for coefficients

R: Number of roots

Returns

TPSCIstate

TPSChem.TPSCIstate — Method

TPSCIstate(clusters::Vector{MOCluster}, fconfig::FockConfig{N}; T=Float64, R=1) where {N}

Constructor using only a single FockConfig. This allows us to turn the CMF state into a TPSCIstate.

Arguments

clusters: vector of clusters types

fconfig: starting FockConfig

T: Type of data for coefficients

R: Number of roots

Returns

TPSCIstate

TPSChem.TPSCIstate — Method

function TPSCIstate(v::TPSCIstate{T,N,R}; T=T, R=R) where {T,N,R}

Constructor creating a TPSCIstate with the same basis as v, but with potentially different R and T. Coefficients of new vector are 0.0

Arguments

T: Type of data for coefficients

R: Number of roots

Returns

TPSCIstate

TPSChem.SPTstate — Type

Represents a state in an set of abitrary (yet low-rank) subspaces of a set of FockConfigs. e.g.

v[FockConfig][TuckerConfig] => Tucker Decomposed Tensor

Data

clusters::Vector{MOCluster}

data::OrderedDict{FockConfig,OrderedDict{TuckerConfig,Tucker}}

p_spaces::Vector{ClusterSubspace}

q_spaces::Vector{ClusterSubspace}

TPSChem.SPTstate — Method

SPTstate(clusters::Vector{MOCluster}, 
    p_spaces::Vector{TPSChem.ClusterSubspace}, 
    q_spaces::Vector{TPSChem.ClusterSubspace})

Constructor - specify input p and q spaces

Arguments

clusters: vector of clusters types

p_spaces: list of p space ranges for each cluster

q_spaces: list of q space ranges for each cluster

Returns

SPTstate

TPSChem.SPTstate — Method

SPTstate(clusters::Vector{MOCluster}, 
    fconfig::FockConfig{N}, 
    cluster_bases::Vector{ClusterBasis}) where {N}

Constructor using only a single FockConfig. This allows us to turn the CMF state into a SPTstate. As such, it chooses the ground state of each cluster in the Fock sector specified by FockConfig to be the P space, and then the Q space is defined as the orthogonal complement of this state within the available basis, specified by cluster_bases.

Arguments

clusters: vector of clusters types

fconfig: starting FockConfig

cluster_basis: list of ClusterBasis types - needed to know the dimensions of the q-spaces

Returns

SPTstate

TPSChem.SPTstate — Method

SPTstate(clusters::Vector{MOCluster}, 
         p_spaces::Vector{ClusterSubspace},
         cluster_bases::Vector{ClusterBasis}; R=1) where {N}

Constructor needing only prespecified p-spaces. The Q space is defined as the orthogonal complement of p-space within the available basis, specified by cluster_bases.

Arguments

clusters: vector of clusters types

p_spaces: list of pspaces for each cluster

cluster_basis: list of ClusterBasis types - needed to know the dimensions of the q-spaces

R: number of roots

Returns

SPTstate

TPSChem.SPTstate — Method

Constructor - create copy, of a particular root

Arguments

v: input SPTstate object

T: data type of new state

R: specific root in new state

Returns

SPTstate

TPSChem.SPTstate — Method

Constructor - create copy, changing T and R optionally

Arguments

v: input SPTstate object

T: data type of new state

R: number of roots in new state

Returns

SPTstate

TPSChem.SPTstate — Method

function SPTstate(ts::BSstate{T,N,R}; thresh=-1, max_number=nothing, verbose=0) where {T,N,R}

Create a SPTstate from a BSstate

Arguments

ts::BSstate

thresh=-1: discard singular values smaller than thresh

max_number=nothing: if != nothing, only keep up to max_number singular vectors per SVD

verbose=0: print level

Returns

SPTstate

TPSChem.FockConfig — Type

config::NTuple{N,Tuple{Int16,Int16}}

Indexes a 'Change in fock space'. For instance, if F1 and F2 are distince FockConfig instances, then TransferConfig is F1-F2.

e.g., config = ((2,3), (2,2), (3,2)) is a 3 cluster configuration with 2, 2, 3 α and 3, 2, 2 β electrons, respectively.

TPSChem.ClusterConfig — Type

config::NTuple{N,Int16}

Indexes an N dimensional space

TPSChem.TuckerConfig — Type

config::NTuple{N,UnitRange{Int}}

Indexes a particular subspace in the tensor product space. Each TuckerConfig instance specifies the entire subspace defined by a set of cluster states (specified by a range) on each cluster.

E.g.,

((1:1), (1:1), (1:1)) is simply the ground state CMF state for a 3 cluster systems.

((2:20), (1:1), (2:20)) are all states where clusters 1 and 2 are excited out of their repsective ground states.

TPSChem.TransferConfig — Type

config::NTuple{N,Tuple{Int16,Int16}}

Indexes a 'Change in fock space'. For instance, if F1 and F2 are distince FockConfig instances, then TransferConfig is F1-F2.

This is mainly used to label individual parts of an operator (Hamiltonian) as they can only enact certain changes in the Fock space configurations.

TPSChem.OperatorConfig — Type

config::Tuple{FockConfig{N}, FockConfig{N}, T, T}

TPSChem.SparseIndex — Type

Abstract index type

TPSChem.ClusterBasis — Type

cluster::MOCluster                            # Cluster to which basis belongs
basis::Dict{Tuple,Matrix{T}}                # Basis vectors (nα, nβ)=>[I,s]

These basis coefficients map local slater determinants to local vectors (nα, nβ): V[αstring*βstring, cluster_state]

TPSChem.ClusterBasis — Method

ClusterBasis(cb::ClusterBasis, T::Type)

Convert from one data type to another

TPSChem.ClusterSubspace — Type

Defines a single cluster's subspace for Tucker. Each focksector is allowed to have a distinct cluster state range for the subspace.

cluster::MOCluster
data::OrderedDict{Tuple{UInt8,UInt8}, UnitRange{Int}}

TPSChem.ClusterOps — Method

ClusterOps(co::ClusterOps, T::Type)

Convert from one data type to another

Methods

Base.setindex! — Method

setindex!(s::AbstractState, a::OrderedDict, b)

TPSChem.add_fockconfig! — Method

add_fockconfig!(s::AbstractState, fock::Vector{Tuple{T,T}}) where T<:Integer

Base.:+ — Method

+(ts1::TPSChem.BSstate, ts2::TPSChem.BSstate)

Base.Multimedia.display — Method

Base.display(s::BSstate; thresh=1e-3)

Pretty print

Base.length — Method

Base.length(s::BSstate)

Base.size — Method

function Base.size(s::BSstate{T,N,R}) where {T,N,R}

LinearAlgebra.dot — Method

dot(ts1::BSstate{T,N,R}, ts2::BSstate{T,N,R}) where {T,N,R}

Returns vector of dot products between each of the states

TPSChem.add! — Method

add!(ts1::TPSChem.BSstate, ts2::TPSChem.BSstate)

Add coeffs in ts2 to ts1

TPSChem.add_config! — Method

TPSChem.add_fockconfig! — Method

function add_fockconfig!(s::BSstate{T,N}, fock::FockConfig) where {T,N}

TPSChem.expand_each_fock_space! — Method

expand_each_fock_space!(s::BSstate, bases)

For each fock space sector defined, add all possible basis states

bases::Vector{ClusterBasis}

TPSChem.eye! — Method

eye!(s::BSstate)

TPSChem.fold! — Method

fold!(ts::BSstate)

TPSChem.get_vector — Method

get_vector(ts::BSstate{T,N,R}, root::Integer) where {T,N,R}

Return a matrix of the core tensors.

TPSChem.get_vector — Method

get_vector(ts::BSstate{T,N,R}) where {T,N,R}

Return a matrix of the core tensors.

TPSChem.mult! — Method

mult!(ts::BSstate, A)

Multiple ts by a matrix A. This is a multiplication over global state index

TPSChem.orthogonalize! — Method

orthogonalize!(ts::BSstate)

Symmetric Orthogonalization of vectors in ts

TPSChem.orthonormalize! — Method

function orthonormalize!(s::BSstate{T,N,R}) where {T,N,R}

orthonormalize

TPSChem.overlap — Method

overlap(ts1::TPSChem.BSstate, ts2::TPSChem.BSstate)

Overlap product between in ts2 to ts1 Returns matrix of overlaps

TPSChem.print_fock_occupations — Method

print_fock_occupations(s::BSstate; root=1, thresh=1e-3)

Pretty print

TPSChem.prune_empty_fock_spaces! — Method

prune_empty_fock_spaces!(s::BSstate)

remove fock_spaces that don't have any configurations

TPSChem.randomize! — Method

randomize!(ts::BSstate; scale=1)

Add some random noise to the vector

TPSChem.scale! — Method

Scale ts by a constant a

TPSChem.set_vector! — Method

set_vector!(ts::BSstate{T,N,R}, v::Matrix{T}) where {T,N,R}

TPSChem.set_vector! — Method

function set_vector!(ts::BSstate{T,N,R}, v::Vector{T}; root=1) where {T,N,R}

TPSChem.unfold! — Method

unfold!(ts::BSstate)

TPSChem.zero! — Method

zero!(s::BSstate)

Base.:* — Method

Base.:*(A::TPSCIstate{T,N,R}, C::AbstractArray) where {T,N,R}

TBW

Base.Multimedia.display — Method

Base.display(s::TPSCIstate; thresh=1e-3, root=1)

Pretty print

LinearAlgebra.dot — Method

dot(v1::TPSCIstate,v2::TPSCIstate; r1=1, r2=1)

LinearAlgebra.dot — Method

dot(v1::TPSCIstate,v2::TPSCIstate; r1=1, r2=1)

LinearAlgebra.norm — Method

norm(s::TPSCIstate, root)

LinearAlgebra.norm — Method

norm(s::TPSCIstate{T,N,R}) where {T,N,R}

TPSChem.add! — Method

add!(s1::TPSCIstate, s2::TPSCIstate)

Add coeffs in s2 to s1

TPSChem.add_fockconfig! — Method

add_fockconfig!(s::TPSCIstate, fock::FockConfig)

TPSChem.add_spin_focksectors — Method

add_spin_focksectors(state::TPSCIstate{T,N,R}) where {T,N,R}

Add the focksectors needed to spin adapt the given TPSCIstate

TPSChem.clip! — Method

clip!(s::TPSCIstate; thresh=1e-5)

TPSChem.ct_analysis — Method

ct_analysis(s::TPSCIstate; ne_cluster=10, thresh=1e-5, nroots=1)

Analyzes charge transfer for each root of the TPSCIstate Prints total weight of charge transfer in each root Only works currently if all clusters have same # of electrons!!

Arguments

s::TPSCIstate

ne_cluster: Int, number of total electrons in each cluster

thresh: Threshold for printing but does not effect total ct

nroots: Total number of roots

TPSChem.ct_table — Method

ct_table(s::TPSCIstate; ne_cluster=10, nroots=1)

Prints total weight of charge transfer in each root in table formate

Arguments

s::TPSCIstate

ne_cluster: Int, number of total electrons in each cluster

nroots: Total number of roots

TPSChem.extract_chosen_root — Method

extract_chosen_root(v::TPSCIstate{T,N,R}, root::Int) -> TPSCIstate{T,N,1}

Return a single-root TPSCIstate containing the coefficients of root from v.

TPSChem.extract_roots — Method

function extract_roots(v::TPSCIstate{T,N,R}, roots)

Extract roots to give new TPSCIstate

TPSChem.eye! — Method

eye!(s::TPSCIstate)

TPSChem.general_ct_analysis — Method

general_ct_analysis(s::TPSCIstate; ne_ref=nothing, thresh=1e-5, nroots=1)

General CT analysis allowing different electron counts per cluster. ne_ref is a Vector{Int} with the reference total electrons (α+β) in each cluster. If ne_ref === nothing, it is taken from the reference Fock space of s.

TPSChem.general_ct_table — Method

general_ct_table(s::TPSCIstate; ne_ref=nothing, nroots=1)

Total CT weight per root, allowing different electrons per cluster. ne_ref is a Vector{Int} with reference total electrons (α+β) in each cluster.

TPSChem.get_vector! — Method

get_vector!(v, s::TPSCIstate)

Fill a preallocated array with the coefficients

TPSChem.get_vector — Method

get_vector(s::TPSCIstate, root=1)

Return a vector of the variables for root. Note that this is the core tensors being returned

TPSChem.get_vector — Method

get_vector(s::TPSCIstate)

Return a matrix of the variables for root.

TPSChem.mult! — Method

mult!(A::TPSCIstate{T,N,R}, C::AbstractArray) where {T,N,R}

TBW

TPSChem.normalize! — Method

normalize!(s::AbstractState)

TPSChem.orth! — Method

orth!(v1::TPSCIstate{T,N,R}) where {T,N,R}

TPSChem.orth_dot — Method

orth_dot(v1::TPSCIstate{T,N,R}, v2::TPSCIstate{T,N,R}) where {T,N,R}

Root-wise dot product between v1 and v2.

Returns the diagonal of overlap(v1, v2), i.e. one dot product per root.

TPSChem.orthonormalize! — Method

function orthonormalize!(s::TPSCIstate{T,N,R}) where {T,N,R}

orthonormalize

TPSChem.overlap — Method

overlap(v1::TPSCIstate{T,N,R}, v2::TPSCIstate{T,N,R}) where {T,N,R}

Compute overlap matrix between v1 and v2

TPSChem.print_configs — Method

print_configs(s::ClusterState; thresh=1e-3)

Pretty print

TPSChem.prune_empty_fock_spaces! — Method

prune_empty_fock_spaces!(s::TPSCIstate)

remove fock_spaces that don't have any configurations

TPSChem.randomize! — Method

function randomize!(s::TPSCIstate{T,N,R}) where {T,N,R}

set all elements to random values, and orthogonalize

TPSChem.scale! — Method

scale!(s::TPSCIstate,c)

TPSChem.set_vector! — Method

function set_vector!(ts::TPSCIstate{T,N,R}, v::Matrix{T}) where {T,N,R}

Fill the coefficients of ts with the values in v

TPSChem.set_vector! — Method

function set_vector!(ts::TPSCIstate{T,N,R}, v::Vector{T}; root=1) where {T,N,R}

Fill the coefficients of ts with the values in v

TPSChem.zero! — Method

zero!(s::TPSCIstate)

set all elements to zero

Base.:* — Method

Base.:*(A::SPTstate{T,N,R}, C::AbstractArray) where {T,N,R}

TBW

Base.:* — Method

Base.:*(A::SPTstate{T,N,R}, C::T) where {T,N,R}

TBW

Base.:+ — Method

Base.:+(A::SPTstate{T,N,R}, B::SPTstate{T,N,R}) where {T,N,R}

TBW

Base.:- — Method

Base.:-(A::SPTstate{T,N,R}, B::SPTstate{T,N,R}) where {T,N,R}

TBW

Base.Multimedia.display — Method

Base.display(s::SPTstate; thresh=1e-3)

Pretty print

Base.length — Method

Base.length(s::SPTstate)

TPSChem.add_fockconfig! — Method

add_fockconfig!(s::SPTstate, fock::FockConfig)

Add an uninitialized FockConfig to the current basis. Typically, you'll want to initialize it right after with add_tuckerconfig!

TPSChem.add_spin_focksectors — Method

add_spin_focksectors(state::SPTstate{T,N,R}) where {T,N,R}

Add the focksectors needed to spin adapt the given SPTstate

TPSChem.compress — Method

compress(ts::SPTstate{T,N,R}; thresh=-1, max_number=nothing, verbose=0) where {T,N,R}

Compress state via HOSVD

Arguments

ts::SPTstate

thresh = -1: threshold for compression

max_number: only keep certain number of vectors per TuckerConfig

verbose=0: print level

Returns

SPTstate

TPSChem.compress_iteratively — Method

compress_iteratively(v::SPTstate, thresh; maxiter=20, verbose=1)

TBW

TPSChem.ct_table — Method

ct_table(s::SPTstate; ne_cluster=10, nroots=1)

Prints total weight of charge transfer in each root in table formate

Arguments

s::SPTstate

ne_cluster: Int, number of total electrons in each cluster

nroots: Total number of roots

TPSChem.eye! — Method

eye!(s::SPTstate)

TPSChem.get_vector — Method

function get_vector(ts::SPTstate{T,N,R}, root::Integer) where {T,N,R}

Return a vector of the variables for root. Note that this is the core tensors being returned

TPSChem.get_vector — Method

function get_vector(ts::SPTstate{T,N,R}) where {T,N,R}

Return a matrix of the core tensors.

TPSChem.nonorth_add! — Method

nonorth_add!(ts1::SPTstate, ts2::SPTstate; thresh=1e-10)

Add coeffs in ts2 to ts1

Note: this does not assume t1 and t2 have the same compression vectors

TPSChem.nonorth_add — Method

nonorth_add(ts1::SPTstate, ts2::SPTstate; thresh=1e-10)

TBW

TPSChem.nonorth_dot — Method

nonorth_dot(ts1::TPSChem.SPTstate, ts2::TPSChem.SPTstate; verbose=0)

Dot product between 1ts2andts1` where each have their own Tucker factors

TPSChem.nonorth_overlap — Method

nonorth_overlap(ts1::TPSChem.SPTstate, ts2::TPSChem.SPTstate; verbose=0)

Dot product between 1ts2andts1` where each have their own Tucker factors

TPSChem.orth_add! — Method

orth_add!(ts1::SPTstate, ts2::SPTstate)

Add coeffs in ts2 to ts1

Note: this assumes t1 and t2 have the same compression vectors

TPSChem.orth_add — Method

orth_add(ts1::SPTstate, ts2::SPTstate)

TBW

TPSChem.orth_dot — Method

orth_dot(ts1::TPSChem.SPTstate, ts2::TPSChem.SPTstate)

Dot product between ts2 and ts1

Warning: this assumes both ts1 and ts2 have the same tucker factors for each TuckerConfig Returns vector of dot products

TPSChem.orth_overlap — Method

orth_overlap(ts1::TPSChem.SPTstate, ts2::TPSChem.SPTstate)

Overlap between ts2 and ts1

This assumes both ts1 and ts2 have the same tucker factors for each TuckerConfig Returns matrix of overlaps

TPSChem.orthonormalize! — Method

function orthonormalize!(s::SPTstate{T,N,R}) where {T,N,R}

orthonormalize

TPSChem.print_fock_occupations — Method

print_fock_occupations(s::SPTstate; thresh=1e-3)

Pretty print

TPSChem.project_into_new_basis — Method

project_into_new_basis(v1::SPTstate{T,N,R}, v2::SPTstate{T,N,R}) where {T,N,R}

Project state v1 into the basis defined by v2

TPSChem.prune_empty_TuckerConfigs! — Method

prune_empty_TuckerConfigs!(s::T) where T<:Union{BSstate, SPTstate}

remove fock_spaces that don't have any configurations

TPSChem.prune_empty_fock_spaces! — Method

prune_empty_fock_spaces!(s::AbstractState)

remove fock_spaces that don't have any configurations

TPSChem.randomize! — Method

TPSChem.scale! — Method

scale!(ts::TPSChem.SPTstate, a::T<:Number)

Scale ts by a constant

TPSChem.scale! — Method

function scale!(ts::TPSChem.SPTstate{T,N,R}, a::Vector{T})

Scale ts by a constant for each state,R

TPSChem.scale — Method

scale(ts::SPTstate{T,N,R}, a::Vector{T}) where {T<:Number, N,R}

TBW

TPSChem.set_vector! — Method

function set_vector!(ts::SPTstate{T,N,R}, v::Matrix{T}) where {T,N,R}

TPSChem.set_vector! — Method

function set_vector!(ts::SPTstate{T,N,R}, v::Vector{T}, root::Int=1) where {T,N,R}

TPSChem.zero! — Method

zero!(s::SPTstate)

Base.convert — Method

function Base.convert(::Type{FockConfig{N}}, in::Vector) where {N}

TPSChem.dim — Method

dim(fc::FockConfig, no)

Return total dimension of space indexed by fc on no orbitals

TPSChem.n_elec — Method

n_elec(fc::FockConfig)

Return number of electrons in fc

TPSChem.n_elec_a — Method

n_elec_a(fc::FockConfig)

Return number of alpha electrons in fc

TPSChem.n_elec_b — Method

n_elec_b(fc::FockConfig)

Return number of beta electrons in fc

TPSChem.possible_spin_focksectors — Method

possible_spin_focksectors(clusters, ref_fock; verbose=1)

Generate list of focksectors needed to spin-complete ref_fock

TPSChem.replace — Method

function replace(tc::FockConfig, idx, fock)

TPSChem.replace — Method

function replace(cc::ClusterConfig{N}, idx, conf) where N

Base.convert — Method

function Base.convert(::Type{TuckerConfig{N}}, in::Vector{UnitRange{T}}) where {T,N}

TPSChem.dim — Method

dim(tc::TuckerConfig)

Return total dimension of space indexed by tc

TPSChem.is_subset — Method

Check if tc1 is a subset of tc2

TPSChem.replace — Method

function replace(cc::TuckerConfig{N}, idx, conf) where N

Base.convert — Method

function Base.convert(::Type{TransferConfig{N}}, in::NTuple{M,Tuple{T,T}}) where {T,N,M}

Base.convert — Method

function Base.convert(::Type{TransferConfig{N}}, in::Vector{Tuple{T,T}}) where {T,N}

TPSChem.replace — Method

function replace(tc::TransferConfig, idx, fock)

TPSChem.merge_cluster_bases — Method

Merge two Vector{ClusterBasis}: for each cluster, union the Fock sectors. If a sector exists in both, keep the one with more roots (columns in vectors). If you want to combine states from both (e.g. augment ground with CT states), pass augment=true — this hcats vectors and re-orthogonalizes via QR.

TPSChem.rotate! — Method

rotate!(cb::ClusterBasis, U::Dict{Tuple,Matrix{T}}) where {T}

Rotate cb by unitary matrices in U

TPSChem.rotate! — Method

rotate!(ops::ClusterOps, U::Dict{Tuple,Matrix{T}}) where {T}

Rotate ops by unitary matrices in U

Base.:+ — Method

Base.:+(a::FockConfig, b::TransferConfig)

Add a FockConfig to a TransferConfig to get a new FockConfig

Base.:- — Method

Base.:-(a::FockConfig, b::FockConfig)

Subtract two FockConfig's, returning a TransferConfig

Base.:- — Method

Base.:-(a::FockConfig, b::TransferConfig)

TPSChem._add_cmf_operator_for_cluster! — Method

function add_cmf_operators!(ops::Vector{ClusterOps}, bases::Vector{ClusterBasis}, ints, Da, Db; verbose=0)

Add effective local hamiltonians (local CASCI) type hamiltonians to a ClusterOps type for each `Cluster'

TPSChem._compute_cluster_ops_for_cluster — Method

compute_cluster_ops(cluster_bases::Vector{ClusterBasis}, ints)

Build the standard local operator tables used by TPSCI. This intentionally does not build the expensive local spin-free 2-RDM table Ppqsr; use compute_cluster_ops_2rdm when calling compute_2rdm.

TPSChem.add_spinfree_2rdm_ops! — Method

add_spinfree_2rdm_ops!(cluster_ops, cluster_bases)

Add the exact local spin-free 2-RDM transition table Ppqsr to an existing cluster_ops object. This is only needed for physically complete compute_2rdm calculations.

TPSChem.compute_cluster_eigenbasis — Method

compute_cluster_eigenbasis(ints::InCoreInts, clusters::Vector{MOCluster}; 
    init_fspace=nothing, delta_elec=nothing, verbose=0, max_roots=10, 
    rdm1a=nothing, rdm1b=nothing, T::Type=Float64)

Return a Vector of ClusterBasis for each Cluster

ints::InCoreInts: In-core integrals

clusters::Vector{MOCluster}: Clusters

verbose::Int: Print level

init_fspace: list of pairs of (nα,nβ) for each cluster for defining reference space for selecting out only certain fock sectors

delta_elec: number of electrons different from reference (init_fspace)

max_roots::Int: Maximum number of vectors for each focksector basis

rdm1a: background density matrix for embedding local hamiltonian (alpha)

rdm1b: background density matrix for embedding local hamiltonian (beta)

ansatze: should be a list of Ansatz objects so that we know how to solve each cluster. Default is FCIAnsatz

T: Data type of the eigenvectors

TPSChem.compute_cluster_eigenbasis_spin — Method

compute_cluster_eigenbasis_spin(   ints::InCoreInts{T}, 
                                   clusters::Vector{MOCluster}, 
                                   rdm1::RDM1{T},
                                   delta_elec::Vector,
                                   ref_fock::FockConfig; 
                                   verbose=0, 
                                   max_roots=10, 
                                   A::Type=FCIAnsatz) where T

Return a Vector of ClusterBasis for each Cluster. For each number of electrons specified by reffock +- 1->deltaelec (for each cluster), we solve the CASCI problem, collecting max_roots of the lowest energy eigenvectors for the half-filled (or of odd number nalpha = nbeta+1) level. Then we apply S^- and S^+ to generate the higher/lower m_s blocks directly.

Arguments

ints: InCoreInts integrals

clusters: Clusters

verbose: Print level

ref_fock: reference space for defining target focksectors with delta_elec

delta_elec: number of electrons different from reference (init_fspace) for each cluster

max_roots::Int: Maximum number of vectors for each focksector basis

rdm1: background density matrix for embedding local hamiltonian

A: the type of Ansatz object used to solve each cluster. Default is FCIAnsatz

T: Data type of the eigenvectors

TPSChem.compute_cluster_est_basis — Method

compute_cluster_est_basis(ints::InCoreInts, clusters::Vector{MOCluster}; 
    init_fspace=nothing, delta_elec=nothing, verbose=0, max_roots=10, 
    rdm1a=nothing, rdm1b=nothing)

Return a Vector of ClusterBasis for each Cluster using the Embedded Schmidt Truncation

ints::InCoreInts: In-core integrals

clusters::Vector{MOCluster}: Clusters

Da: background density matrix for embedding local hamiltonian (alpha)

Db: background density matrix for embedding local hamiltonian (beta)

init_fspace: list of pairs of (nα,nβ) for each cluster for defining reference space for selecting out only certain fock sectors

thresh_schmidt: the threshold for the EST

thresh_orb: threshold for the orbital

thresh_ci: threshold for the ci problem

TPSChem.compute_cluster_ops_2rdm — Method

compute_cluster_ops_2rdm(cluster_bases, ints)

Build the standard TPSCI local operators plus the expensive Ppqsr table required by compute_2rdm. Use this instead of compute_cluster_ops only when you need a 2-RDM.

TPSChem.form_schmidt_basis — Method

form_schmidt_basis

threshorb : threshold for determining how many bath orbitals to include threshschmidt : threshold for determining how many singular vectors to include for cluster basis

Returns new basis for the cluster

TPSChem.get_ortho_compliment — Method

get_ortho_compliment(tss::ClusterSubspace, cb::ClusterBasis)

For a given ClusterSubspace, tss, return the subspace remaining

TPSChem.tdm_A — Method

tdm_A(cb::ClusterBasis; verbose=0)

Compute <s|p'|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

Returns Dict[((na,nb),(na,nb))] => Array

TPSChem.tdm_AA — Method

tdm_AA(cb::ClusterBasis; verbose=0)

Compute <s|p'q'|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

Returns Dict[((na,nb),(na,nb))] => Array

TPSChem.tdm_AAa — Method

tdm_AAa(cb::ClusterBasis, spin_case; verbose=0)

Compute <s|p'q'r|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

spin_case: alpha or beta

Returns Dict[((na,nb),(na,nb))] => Array

TPSChem.tdm_AB — Method

tdm_AB(cb::ClusterBasis; verbose=0)

Compute <s|p'q'|t> between all cluster states, s and t from accessible sectors of a cluster's fock space, where p' is alpha and q' is beta.

Returns Dict[((na,nb),(na,nb))] => Array

TPSChem.tdm_ABa — Method

tdm_ABa(cb::ClusterBasis, spin_case; verbose=0)

Compute <s|p'q'r|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

spin_case: alpha or beta

Returns Dict[((na,nb),(na,nb))] => Array

TPSChem.tdm_Aa — Method

tdm_Aa(cb::ClusterBasis, spin_case; verbose=0)

Compute <s|p'q|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

spin_case: alpha or beta

Returns Dict[((na,nb),(na,nb))] => Array

TPSChem.tdm_Ab — Method

tdm_Ab(cb::ClusterBasis; verbose=0)

Compute <s|p'q|t> between all cluster states, s and t from accessible sectors of a cluster's fock space, where p' is alpha and q is beta.

Returns Dict[((na,nb),(na,nb))] => Array

TPSChem.tdm_H — Method

tdm_H(cb::ClusterBasis; verbose=0)

Compute local Hamiltonian <s|H|t> between all cluster states, s and t from accessible sectors of a cluster's fock space.

Returns Dict[((na,nb),(na,nb))] => Array

TPSChem.tdm_S2 — Method

TPSChem.tdm_spinfree_2rdm — Method

tdm_spinfree_2rdm(cb::ClusterBasis; verbose=0)

Compute the local spin-free two-body transition density

Gamma[p,q,r,s,u,v] = sum_{sigma,tau} <u|p'_sigma q'_tau s_tau r_sigma|v>

for each Fock-neutral sector of a cluster basis. The first four dimensions are local orbital indices in the same convention as compute_2rdm; the final two dimensions are bra and ket cluster-basis state indices.