CMF

Background

The Cluster Mean Field (CMF) approach used in this work was developed by Jimènez-Hoyos and Scuseria.

PySCF-backed helpers (pyscf_do_scf, pyscf_build_ints, ...) are provided by the package extension TPSChemPyCallExt, which activates when PyCall is loaded.

Index

TPSChem.ClusterMeanField.assemble_full_rdm

TPSChem.ClusterMeanField.assemble_full_rdm

TPSChem.ClusterMeanField.cmf_ci

TPSChem.ClusterMeanField.cmf_ci

TPSChem.ClusterMeanField.cmf_ci_iteration

TPSChem.ClusterMeanField.cmf_ci_iteration

TPSChem.ClusterMeanField.cmf_oo

TPSChem.ClusterMeanField.cmf_oo

TPSChem.ClusterMeanField.cmf_oo_diis

TPSChem.ClusterMeanField.cmf_oo_gd

TPSChem.ClusterMeanField.compute_cmf_energy

TPSChem.ClusterMeanField.orbital_gradient_numerical

TPSChem.ClusterMeanField.orbital_objective_function

TPSChem.ClusterMeanField.pack_gradient

TPSChem.ClusterMeanField.unpack_gradient

TPSChem.QCBase.compute_energy

Documentation

TPSChem.ClusterMeanField.assemble_full_rdm — Method

assemble_full_rdm(clusters::Vector{MOCluster}, rdm1s::Dict{Integer, Array}, rdm2s::Dict{Integer, Array})

Return full system 1 and 2 RDMs

TPSChem.ClusterMeanField.assemble_full_rdm — Method

assemble_full_rdm(clusters::Vector{MOCluster}, rdm1s::Dict{Integer, RDM1{T}}) where T

Return spin summed 1 and 2 RDMs

TPSChem.ClusterMeanField.cmf_ci — Method

cmf_ci(ints, clusters, fspace, in_rdm1::RDM1; 
            maxiter_ci  = 100, 
            maxiter_d1  = 20, 
            tol_d1      = 1e-6, 
            tol_ci      = 1e-8, 
            verbose     = 1,
            sequential  = false)

Optimize the 1RDM for CMF-CI

#Arguments

ints::InCoreInts: integrals for full system

clusters::Vector{MOCluster}: vector of cluster objects

fspace::Vector{Vector{Integer}}: vector of particle number occupations for each cluster specifying the sectors of fock space

in_rdm1: initial guess for 1particle density matrix

tol_d1: Convergence threshold for change in density

tol_ci: Convergence threshold for the cluster CI problems

maxiter_d1: Max number of iterations for density optimization

maxiter_ci: Max number of iterations for CI diagonalization

sequential: Use the density matrix of the previous cluster in a cMF iteration to form effective integrals. Improves comvergence, may depend on cluster orderings

verbose: Printing level

Returns

e: Energy

rdm1_dict: Dictionary of 1RDMs; cluster index –> RDM1

rdm2_dict: Dictionary of 2RDMs; cluster index –> RDM2

TPSChem.ClusterMeanField.cmf_ci — Method

cmf_ci(mol::Molecule, C::Matrix, clusters::Vector{MOCluster}, fspace::Vector, dguess; 
        max_iter=10, dconv=1e-6, econv=1e-10, verbose=1)

Optimize the 1RDM for CMF-CI, without requiring an InCoreInts object

Arguments

mol::Molecule: a Molecule type

C: MO coefficients for full system (spin restricted)

clusters::Vector{MOCluster}: vector of MOCluster objects

fspace::Vector{Vector{Integer}}: vector of particle number occupations for each cluster specifying the sectors of fock space

dguess: initial guess for 1particle density matrix (spin summed)

dconv: Convergence threshold for change in density

econv: Convergence threshold for change in energy

verbose: how much to print

TPSChem.ClusterMeanField.cmf_ci_iteration — Method

cmf_ci_iteration(mol::Molecule, C, rdm1a, rdm1b, clusters, fspace; 
                 verbose=1)

Perform single CMF-CI iteration, returning new energy, and density. This method forms the eri's on the fly to avoid global N^4 storage

Arguments

mol::Molecule: a Molecule type

C: MO coefficients for full system (spin restricted)

rdm1a: 1particle density matrix (alpha)

rdm1b: 1particle density matrix (beta)

clusters::Vector{MOCluster}: vector of MOCluster objects

fspace::Vector{Vector{Int}}: vector of particle number occupations for each cluster specifying the sectors of fock space

verbose: how much to print

ci_max_iter: How many FCI iterations do we allow for each cluster. Default=40

ci_conv_tol: How what is the convergence (gradient norm) for the FCI on each cluster. Default=1e-8

See also: cmf_ci_iteration

TPSChem.ClusterMeanField.cmf_ci_iteration — Method

cmf_ci_iteration(ints::InCoreInts{T}, clusters::Vector{MOCluster}, in_rdm1::RDM1{T}, fspace; 
                      use_pyscf = false, 
                      verbose   = 1, 
                      sequential= false, 
                      spin_avg  = true, 
                      tol_ci    = 1e-8,
                      maxiter_ci    = 100
) where T

Perform single CMF-CI iteration, returning new energy, and density

TPSChem.ClusterMeanField.cmf_oo — Method

cmf_oo(mol::Molecule, Cguess::Matrix, clusters::Vector{MOCluster}, fspace, dguess; 
        max_iter_oo=100, max_iter_ci=100, gconv=1e-6, verbose=0, method="bfgs")

Do CMF with orbital optimization with on the fly integrals

TPSChem.ClusterMeanField.cmf_oo — Method

cmf_oo(ints::InCoreInts{T}, clusters::Vector{MOCluster}, fspace, dguess::RDM1{T}; 
            max_iter_oo=100, 
            max_iter_ci=100, 
            gconv=1e-6, 
            verbose=0, 
            method="bfgs", 
            alpha=nothing,
            sequential=false) where T

Do CMF with orbital optimization

#Arguments

ints::InCoreInts: integrals for full system

clusters::Vector{MOCluster}: vector of cluster objects

fspace::Vector{Vector{Integer}}: vector of particle number occupations for each cluster specifying the sectors of fock space

dguess_a: initial guess for 1particle density matrix

max_iter_oo: Max iter for the orbital optimization iterations

max_iter_ci: Max iter for the cmf iteration for the cluster states

gconv: Convergence threshold for change in gradient of energy

sequential: If true use the density matrix of the previous cluster in a cMF iteration to form effective integrals. Improves comvergence, may depend on cluster orderings

verbose: Printing level

method: optimization method

TPSChem.ClusterMeanField.cmf_oo_diis — Method

cmf_oo_diis( ints_in::InCoreInts{T}, clusters::Vector{MOCluster}, fspace, dguess::RDM1{T}; 
                maxiter_oo      = 100, 
                maxiter_ci      = 100, 
                maxiter_d1      = 100, 
                tol_oo          = 1e-6, 
                tol_d1          = 1e-7, 
                tol_ci          = 1e-8, 
                verbose         = 0, 
                max_ss_size     = 8, 
                diis_start      = 1,
                alpha           = .1,
                zero_intra_rots = true 
                sequential      = false
                ) where T

Do CMF with orbital optimization using DIIS

Arguments

ints::InCoreInts: integrals for full system

clusters::Vector{MOCluster}: vector of cluster objects

fspace::Vector{Vector{Integer}}: vector of particle number occupations for each cluster specifying the sectors of fock space

dguess: initial guess for 1particle density matrix

maxiter_oo: Max iter for the orbital optimization iterations

maxiter_d1: Max iter for the cmf iteration for the 1RDM

maxiter_ci: Max iter for the CI diagonalization of the cluster states

tol_oo: Convergence threshold for change in orbital gradient

tol_ci: Convergence threshold for the cluster CI problems

tol_d1: Convergence threshold for the CMF 1RDM

sequential: If true use the density matrix of the previous cluster in a cMF iteration to form effective integrals. Improves comvergence, may depend on cluster orderings

verbose: Printing level

max_ss_size: Max number of DIIS vectors

diis_start: When to start doing DIIS extrapolations

alpha: New vector added to ss is k=-alpha*g where g is the orbital gradient. This should be improved with Hessian.

zero_intra_rots: Should we zero out rotations within a cluster? Helps with FCI solvers which should have zero gradients.

Returns

e: Energy

U::Matrix: Orbital rotation matrix from input to output orbitals

d1::RDM1: Optimized 1RDM in the optimized orbital basis

TPSChem.ClusterMeanField.cmf_oo_gd — Method

cmf_oo_gd( ints_in::InCoreInts{T}, clusters::Vector{MOCluster}, fspace, dguess::RDM1{T}; 
                maxiter_oo      = 100, 
                maxiter_ci      = 100, 
                maxiter_d1      = 100, 
                tol_oo          = 1e-6, 
                tol_d1          = 1e-7, 
                tol_ci          = 1e-8, 
                verbose         = 0, 
                alpha           = .1,
                zero_intra_rots = true,
                sequential      = false) where T

Do CMF with orbital optimization

Arguments

ints::InCoreInts: integrals for full system

clusters::Vector{MOCluster}: vector of cluster objects

fspace::Vector{Vector{Integer}}: vector of particle number occupations for each cluster specifying the sectors of fock space

dguess: initial guess for 1particle density matrix

maxiter_oo: Max iter for the orbital optimization iterations

maxiter_d1: Max iter for the cmf iteration for the 1RDM

maxiter_ci: Max iter for the CI diagonalization of the cluster states

tol_oo: Convergence threshold for change in orbital gradient

tol_ci: Convergence threshold for the cluster CI problems

tol_d1: Convergence threshold for the CMF 1RDM

sequential: If true use the density matrix of the previous cluster in a cMF iteration to form effective integrals. Improves comvergence, may depend on cluster orderings

verbose: Printing level

Returns

e: Energy

U::Matrix: Orbital rotation matrix from input to output orbitals

d1::RDM1: Optimized 1RDM in the optimized orbital basis

TPSChem.ClusterMeanField.compute_cmf_energy — Method

compute_cmf_energy(mol::Molecule, C::Matrix, rdm1s, rdm2s, clusters)

Compute the energy of a cluster-wise product state (CMF), specified by a list of 1 and 2 particle rdms local to each cluster

#Arguments

mol::Molecule

C::Matrix: MO coefficients

rdm1s: dictionary (ci.idx => Array) of 1rdms from each cluster (spin summed)

rdm2s: dictionary (ci.idx => Array) of 2rdms from each cluster (spin summed)

clusters::Vector{MOCluster}: vector of cluster objects

return the total CMF energy

TPSChem.ClusterMeanField.orbital_gradient_numerical — Method

orbital_gradient_numerical(ints, clusters, kappa, fspace, da, db; 
                                gconv = 1e-8, 
                                verbose = 1,
                                stepsize = 1e-6)

Compute orbital gradient with finite difference

TPSChem.ClusterMeanField.orbital_objective_function — Method

orbital_objective_function(ints, clusters, kappa, fspace, da, db; 
                                ci_conv     = 1e-9,
                                sequential  = false,
                                verbose     = 1)

Objective function to minimize in OO-CMF

TPSChem.ClusterMeanField.pack_gradient — Method

pack_gradient(K,norb)

TPSChem.ClusterMeanField.unpack_gradient — Method

unpack_gradient(kappa,norb)

TPSChem.QCBase.compute_energy — Method

QCBase.compute_energy(ints::InCoreInts{T}, rdm1s::Dict{Integer,RDM1{T}}, rdm2s::Dict{Integer,RDM2{T}}, clusters::Vector{MOCluster}; verbose=0) where T

Compute the energy of a cluster-wise product state (CMF), specified by a list of 1 and 2 particle rdms local to each cluster. This method uses the full system integrals.

ints::InCoreInts: integrals for full system

rdm1s: dictionary (ci.idx => RDM1) of 1rdms from each cluster

rdm2s: dictionary (ci.idx => RDM2) of 2rdms from each cluster

clusters::Vector{MOCluster}: vector of cluster objects

return the total CMF energy