SPT

The Subspace Product Tucker (SPT) method (formerly called Block-Sparse Tucker, BST) approximates FCI as a linear combination of individually compressed (via HOSVD) blocks of the Hilbert space.

Background

Similar to TPSCI, the approach here starts with a CMF wavefunction, and systematically reintroduces the discarded tensor product states to variationally approach FCI. Unlike with TPSCI, however, we don't assume that the final wavefunction is written as a purely sparse form (where only a few TPS's are needed), but rather we assume that collections of TPS's where certain numbers of clusters are "excited" can be efficiently compressed via HOSVD (although the basic idea would extend to other tensor decompositions, like CP or MPS).

Performance considerations

Index

TPSChem.Tucker

TPSChem.Tucker

Base.:+

TPSChem.add!

TPSChem.add_transformed_tensor

TPSChem.build_compressed_1st_order_state

TPSChem.build_compressed_1st_order_state_old

TPSChem.build_sigma!

TPSChem.build_sigma_cepa!

TPSChem.build_sigma_serial!

TPSChem.cache_hamiltonian_old

TPSChem.ci_solve

TPSChem.compress

TPSChem.compute_expectation_value

TPSChem.compute_spin_residual

TPSChem.define_foi_space

TPSChem.get_map

TPSChem.nonorth_add

TPSChem.nonorth_dot

TPSChem.nonorth_dot

TPSChem.nonorth_overlap

TPSChem.orth_dot

TPSChem.project_out!

TPSChem.scale

TPSChem.scale!

TPSChem.subspace_product_tucker

TPSChem.subspace_product_tucker_old

TPSChem.transform_basis

TPSChem.transform_basis!

TPSChem.tucker_cepa_solve

TPSChem.tucker_cepa_solve2

TPSChem.tucker_decompose

TPSChem.tucker_initialize

TPSChem.tucker_recompose

Documentation

TPSChem.build_sigma! — Method

build_sigma_parallel!(sigma_vector::SPTstate, ci_vector::SPTstate, cluster_ops, clustered_ham)

TPSChem.build_sigma_cepa! — Method

build_sigma_cepa!(sigma_vector, ci_vector, cluster_ops, clustered_ham)

Thread-safe version of build_sigma! for use in CEPA matvec. Uses per-task scratch allocation to avoid races with Julia's dynamic scheduler.

TPSChem.build_sigma_serial! — Method

build_sigma!(sigma_vector::SPTstate, ci_vector::SPTstate, cluster_ops, clustered_ham)

TPSChem.cache_hamiltonian_old — Method

build_sigma!(sigma_vector::SPTstate, ci_vector::SPTstate, cluster_ops, clustered_ham)

TPSChem.build_compressed_1st_order_state — Method

build_compressed_1st_order_state(ψ::SPTstate{T,N,R}, cluster_ops, clustered_ham; 
thresh=1e-7, 
max_number=nothing, 
nbody=4, 
prescreen=false,
compress_twice=true
)  where {T,N,R}

TBW

TPSChem.build_compressed_1st_order_state_old — Method

build_compressed_1st_order_state_old(ket_cts::SPTstate{T,N,R}, cluster_ops, clustered_ham; 
    thresh=1e-7, 
    max_number=nothing, 
    nbody=4, 
    compress_twice=true) where {T,N,R}
    thresh=1e-7, 
    max_number=nothing, 
    nbody=4) where {T,N}

Apply the Hamiltonian to v expanding into the uncompressed space. This is done only partially, where each term is recompressed after being computed. Lots of overhead probably from compression, but never completely uncompresses.

#Arguments

cts::SPTstate: input state

cluster_ops:

clustered_ham: Hamiltonian

thresh: Threshold for each HOSVD

max_number: max number of tucker factors kept in each HOSVD

nbody: allows one to limit (max 4body) terms in the Hamiltonian considered

compress_twice: Should we recompress after adding the tuckers together

#Returns

v1::SPTstate

TPSChem.ci_solve — Method

function ci_solve(ci_vector_in::SPTstate{T,N,R}, cluster_ops, clustered_ham; 
                     conv_thresh = 1e-5,
                     max_ss_vecs = 12,
                     max_iter    = 40,
                     shift       = nothing,
                     precond     = false,
                     verbose     = 0,
                     solver      = "davidson") where {T,N,R}

Solve for ground state in the space spanned by ci_vector's compression vectors

Arguments

conv_thresh: residual convergence threshold

max_ss_vecs: max number of subspace vectors

max_iter: Max iterations in solver

shift: Use a shift? this is for CEPA type

precond: use preconditioner? Only applied to Davidson and not yet working,

verbose: print level

solver: Which solver to use. Options = ["davidson", "krylovkit"]

TPSChem.define_foi_space — Method

define_foi_space(v::SPTstate, clustered_ham; nbody=2)

Compute the first-order interacting space as defined by clustered_ham

#Arguments

v::SPTstate: input state

clustered_ham: Hamiltonian

nbody: allows one to limit (max 4body) terms in the Hamiltonian considered

#Returns

foi::OrderedDict{FockConfig,Vector{TuckerConfig}}

TPSChem.get_map — Method

get_map(ci_vector::SPTstate, cluster_ops, clustered_ham)

Get LinearMap with takes a vector and returns action of H on that vector

TPSChem.project_out! — Method

project_out!(v::SPTstate, w::SPTstate; thresh=1e-16)

Project w out of v |v'> = |v> - |w><w|v>

TPSChem.tucker_cepa_solve — Method

tucker_cepa_solve!(ref_vector::SPTstate, cepa_vector::SPTstate, cluster_ops, clustered_ham; tol=1e-5, cache=true)

Arguments

ref_vector: Input reference state.

cepa_vector: SPTstate which defines the configurational space defining {X}. This

should be the first-order interacting space (or some compressed version of it).

cluster_ops

clustered_ham

tol: haven't yet set this up (NYI)

cache: Should we cache the compressed H operators? Speeds up drastically, but uses lots of memory

Compute compressed CEPA. Since there can be non-zero overlap with a multireference state, we need to generalize.

HC = SCe

|Haa + Hax| |1 | = |I   + Sax| |1 | E
|Hxa + Hxx| |Cx|   |Sxa + I  | |Cx|

Haa + Hax*Cx = (1 + Sax*Cx)E
Hxa + HxxCx = SxaE + CxE

The idea for CEPA is to approximate E in the amplitude equation. CEPA(0): E = Eref

(Hxx-Eref)*Cx = Sxa*Eref - Hxa

Ax=b

After solving, the Energy can be obtained as:

E = (Eref + Hax*Cx) / (1 + Sax*Cx)

TPSChem.tucker_cepa_solve2 — Method

tucker_cepa_solve!(ref_vector::SPTstate, cepa_vector::SPTstate, cluster_ops, clustered_ham; tol=1e-5, cache=true)

Arguments

ref_vector: Input reference state.

cepa_vector: SPTstate which defines the configurational space defining {X}. This

should be the first-order interacting space (or some compressed version of it).

cluster_ops

clustered_ham

tol: haven't yet set this up (NYI)

cache: Should we cache the compressed H operators? Speeds up drastically, but uses lots of memory

Compute compressed CEPA. Since there can be non-zero overlap with a multireference state, we need to generalize.

HC = SCe

|Haa + Hax| |1 | = |I + Sax| |1 | E |Hxa + Hxx| |Cx| |Sxa + I | |Cx|

Haa + HaxCx = (1 + SaxCx)E Hxa + HxxCx = SxaE + CxE

The idea for CEPA is to approximate E in the amplitude equation. CEPA(0): E = Eref

(Hxx-Eref)Cx = SxaEref - Hxa

Ax=b

After solving, the Energy can be obtained as: E = (Eref + HaxCx) / (1 + SaxCx)

TPSChem.compute_expectation_value — Method

function compute_expectation_value(ci_vector::SPTstate{T,N,R}, cluster_ops, clustered_op::TPSChem.ClusteredOperator; nbody) where {T,N,R}

TPSChem.compute_spin_residual — Method

compute_spin_residual(v::SPTstate{T,N,R}, cluster_ops, thresh) where {T,N,R}

Compute |r> = S2|v> - |v><v|S2|v>

TPSChem.subspace_product_tucker — Method

subspace_product_tucker(input_vec::SPTstate{T,N,R}, cluster_ops, clustered_ham;
                    max_iter=20,
                    nbody=4,
                    H0="Hcmf",
                    thresh_var=1e-4,
                    thresh_foi=1e-6,
                    thresh_pt=1e-5,
                    thresh_spin=nothing,
                    ci_conv=1e-5,
                    ci_max_iter=50,
                    ci_max_ss_vecs=12,
                    ci_lindep_thresh=1e-10,
                    resolve_ss=false,
                    do_pt=true,
                    tol_tucker=1e-6,
                    solver="davidson",
                    verbose=1) where {T,N,R}

Arguments

input_vec::SPTstate: initial state

cluster_ops: local cluster operators

clustered_ham::ClusteredOperator: hamiltonian

max_iter = 20: max number of iterations

nbody = 4: include up to nbody terms when forming the FOIS to search

H0 = "Hcmf": zeroth-order hamiltonian for computing 1st order wavefunction ["Hcmf", "H"]

nbody: max number of nbody terms in the Hamiltonian used for creating FOIS

thresh_var: Compression threshold for the variational solution

thresh_foi: Compression threshold for the FOIS

thresh_pt: Compression threshold for the first-order wavefunction (if used)

thresh_spin: Compression threshold for adding S2 residual, only does this if thresh is specified

ci_conv: Convergence threshold for the CI (norm of residual)

ci_max_iter: max iterations for CI problem

ci_max_ss_vecs: max number of subspace vectors for lanczos/davidson

resolve_ss: After compressing previous variational state, should we resolve in new subspace?

do_pt = true: Compute pt1 wavefunction for finding updated compression basis?

tol_tucker: Convergence threshold for Tucker iterations (energy change)

solver:

verbose: How much to print?

Returns

e_var::Float64: the final variational energy

v_var::SPTstate: the final variational state

See also: SPTstate, Tucker

TPSChem.subspace_product_tucker_old — Method

function subspace_product_tucker(input_vec::SPTstate{T,N,R}, cluster_ops, clustered_ham;
    max_iter        = 20,
    max_iter_pt     = 200, # max number of iterations for solving for PT1
    nbody           = 4,
    H0              = "Hcmf",
    thresh_var      = 1e-4,
    thresh_foi      = 1e-6,
    thresh_pt       = 1e-5,
    ci_conv         = 1e-5,
    ci_max_iter     = 50,
    ci_max_ss_vecs  = 12,
resolve_ss      = false,
    do_pt           = true,
    tol_tucker      = 1e-6 ) where {T,N,R}

Arguments

input_vec::SPTstate: initial state

cluster_ops: local cluster operators

clustered_ham::ClusteredOperator: hamiltonian

max_iter = 20: max number of iterations

max_iter_pt = 200: max number of iterations to solve for the 1st order wavefunction

nbody = 4: include up to nbody terms when forming the FOIS to search

H0 = "Hcmf": zeroth-order hamiltonian for computing 1st order wavefunction ["Hcmf", "H"]

nbody: max number of nbody terms in the Hamiltonian used for creating FOIS

thresh_var: Compression threshold for the variational solution

thresh_foi: Compression threshold for the FOIS

thresh_pt: Compression threshold for the first-order wavefunction (if used)

ci_conv: Convergence threshold for the CI (norm of residual)

ci_max_iter: max iterations for CI problem

ci_max_ss_vecs: max number of subspace vectors for lanczos/davidson

resolve_ss: After compressing previous variational state, should we resolve in new subspace?

do_pt = true: Compute pt1 wavefunction for finding updated compression basis?

tol_tucker: Convergence threshold for Tucker iterations (energy change)

solver:

Returns

e_var::Float64: the final variational energy

v_var::SPTstate: the final variational state

See also: SPTstate, Tucker

HOSVD

TPSChem.Tucker — Type

Simple Tucker (HOSVD) type

Data

core::Array{T, N}

factors::NTuple{N, Matrix{T}}

Tucker factors are stored as tall matrices the core tensor is actually a list of core tensors, to enable multistate calculations

TPSChem.Tucker — Method

function Tucker(t::Tucker{T,N,R}, n_roots::Integer) where {T,N,R}

Create a new Tucker with a different number of core tensors; R=n_roots. New core tensors added are initialized to zero.

Base.:+ — Method

Base.:(+)(t1::NTuple{R,Array}, t2::NTuple{R,Array}) where R

TBW

TPSChem.add! — Method

add!(t1::NTuple{R,Array}, t2::NTuple{R,Array}) where R

TBW

TPSChem.add_transformed_tensor — Method

add_transformed_tensor(v::Array{T,N}, transforms::NTuple{N,Matrix{T}}, out::Array{T,N}, scr::Vector{Vector{T}}; trans=false) where {T<:Number,N}

Transform a tensor (v) by a sequence of matrices (transforms). This version takes in a set of scratch arrays (scr), one for each index in the tensor

out(IJK) += v(ijk) U(Ii) U(Jj) U(Kk)

TPSChem.compress — Method

compress(t::Tucker{T,N}; thresh=1e-7, max_number=nothing) where {T,N}

Try to compress further

TPSChem.nonorth_add — Method

nonorth_add(tucks::Vector{Tucker{T,N}}; thresh=1e-10, max_number=nothing) where {T,N}

Add together multiple Tucker instances. Assumed non-orthogonal.

Arguments

tucks::Vector{Tucker{T,N}}: Vector of Tucker objects

TPSChem.nonorth_dot — Method

function nonorth_dot(t1::Tucker{T,N,R}, t2::Tucker{T,N,R}) where {T,N,R}

Note: This doesn't assume t1 and t2 have the same compression vectors

TPSChem.nonorth_dot — Method

function nonorth_dot(t1::Tucker{T,N,R}, t2::Tucker{T,N,R}) where {T,N,R}

Note: This doesn't assume t1 and t2 have the same compression vectors

TPSChem.nonorth_overlap — Method

function nonorth_overlap(t1::Tucker{T,N,R}, t2::Tucker{T,N,R}) where {T,N,R}

Note: This doesn't assume t1 and t2 have the same compression vectors. Returns RxR overlap matrix

TPSChem.orth_dot — Method

function orth_dot(t1::Tucker{T,N,R}, t2::Tucker{T,N,R}) where {T,N,R}

Note: This assumes t1 and t2 have the same compression vectors

TPSChem.scale! — Method

scale!(t1::Tucker, a)

TBW

TPSChem.scale — Method

scale(t1::Tucker, a)

TBW

TPSChem.transform_basis! — Method

transform_basis!(tuck::Tucker{T,N,R}, transform_list::Vector{Matrix{T}}) where {T,N,R}

TBW

TPSChem.transform_basis — Method

transform_basis(v::Array{T,N}, transforms::NTuple{N,Matrix{T}}; trans=false) where {T<:Number,N}

TBW

TPSChem.tucker_decompose — Method

function tucker_decompose(A::Array{T,N}; thresh=1e-7, max_number=nothing, verbose=1, type="magnitude") where {T,N}

Tucker Decomposition of dense tensor: A ~ X *(1) U1 *(2) U2 .... where cluster states are discarded based on the corresponding SVD #Arguments

A: matrix to decompose

thresh: threshold for discarding tucker factors

max_number: limit number of tucker factors to this value

type: type of trunctation. "magnitude" discards values smaller than this number. "sum" discards values such that the sum of discarded values is smaller than thresh.

TPSChem.tucker_initialize — Method

function tucker_decompose(A::Array{T,N}; num_roots=nothing, max_number=nothing, verbose=1, type="magnitude") where {T,N}

Tucker Decomposition of dense tensor: A ~ X *(1) U1 *(2) U2 .... where cluster states are discarded based on the corresponding SVD #Arguments

A: matrix to decompose

num_roots: number of roots to keep

max_number: limit number of tucker factors to this value

TPSChem.tucker_recompose — Method

tucker_recompose(core, factors)

Recompose Tucker Decomposition